refactor(prerender): extract cleanPrerenderedHtml for testability
Move cleanPrerenderedHtml() and META_TAG_PATTERNS from prerender.ts
into a standalone module (scripts/cleanPrerenderedHtml.ts) so the
function can be unit-tested without Playwright dependencies.
